+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/Shared/Serialization/OpenMetricsSerializer.cs
@@ -16,6 +16,8 @@ internal abstract class OpenMetricsSerializer : TextFormatSerializer
 
     protected override string TargetInfoTypeValue => "info";
 
+    protected override bool EscapeHelpQuotationMarks => true;
+
     public override int WriteEof(byte[] buffer, int cursor)
     {
         // OpenMetrics expositions MUST be terminated with "# EOF".